Allow the minimum set of verified data from Verify
The `from` and `to` dates will only exist if a user has changed some of
their details and even then it may not have a `to` date (if the detail
is current)

It's also possible for no middle name to be returned, our code would
error on the `fetch` with an empty middle name. This change will allow
empty middle names.