Remove hint on ITT/QTS year
Originally this was a select with only eligible options so we needed
the hint to explain why their options were limited. Now that it has been
amended to radios and a path to fail has been added this hint text has
become very leading and it is no longer required.