Simplify student loan plan gate question
We ask this question to determine whether we need to know the student
loan repayment plan the claimant is on so we can cover the recoveries
due for their claim. It makes no sense to ask them if they have ever
taken out a student loan at this stage, as if they hadn't they
wouldn't be claiming back student loan repayments!

Separate work might follow to add an additional "Have you ever taken
out a student?" question earlier in the journey to tell those that might
be claiming even though they've never had one that they aren't eligible.