Adjust escaping of formula-triggering characters
Just wrapping the entire value in ="STRING_HERE" is not going to prevent
a formula triggering, for example a value of test"+SUM(A1- A2)+" would
get escaped to ="test"+SUM(A1- A2)+"", which would still attempt to
evaluate the entered info as a formula.

Instead we escape the first character if it is one of the known
formula-triggering characters. This may result in operability issues
further down the line, but these will be easy to deal with by the human
operator that will be working with the files. The excess of backslashes
in the gsub is the only way to add the '\' escape character and match
substitution character '\0'.

See http://georgemauer.net/2017/10/07/csv-injection.html for more
details of how opening CSV files with Excel can result in formula
injection.