Reverse the 50% question to ask about leadership duties
We need to keep `taught_eligible_subjects` so that we can require an
explicit answer to the subjects taught question. We can't look for an
empty array because that would allow users to click through without
making a selection.

The subjects taught question doesn't need to reset the leadership
question any more as they are now decoupled.

This also amends the "no" answer for subjects taught to be grammatically
correct given the question. "No applicable" implies that the question
doesn't apply to the user, not that they don't teach any of the given
subjects.