Import the close date into School
When a school changes some of its attributes (name, status, etc) it is
closed in the Get information about schools data and re-opened with the
new attributes.

This can cause duplicates in the school data we have and show to users.
In order to differentiate the duplicates the close date will be used.

This commit adds the close_date to the model and school_data importer.