Add cookie functions for getting and setting cookies
We need to be able to check if a user has accepted cookies on our
service. In order to do that we need some utility functions that
can get or set cookies.

The intention here is that if we do need  to set a non-essential cookie,
a developer would use the `setNonEssentialCookie` method that will only
set the cookie if consent has been given by the user. This could be a
case of YAGNI, but I felt it important to have this documented.