Closed schools show close date during search
Searching for a claim school or a current school shows the date a school
closed letting the user make an informed decision about which school to
select.

Updates the autocomplete javascript to use the school id instead of the
name.

Orders the school search results by name and then close_date so the
closed school appears beneath on schools with duplicate names.

This is an interim solution until we find the right way to deal with
closed schools.