Verified middle names are now optional
We saw an error on production where a user returned from Verify with an
unverified middle name causing an application error and they couldn't
complete their journey.

Middle names are not essential to payroll so we can mark them as
optional, if a non-verified value comes back we can ignore it.

https://rollbar.com/dxw/teachers-payment-service/items/69/