Create a staging slot for the main app service
We want to be able to deploy to a staging slot so we can have
zero-downtime deployments.

The runtime stack is still declared for the "production" slot so an
incomplete deployment followed by a service restart doesn't take the
service down. If not specified, the fact that we do a `Complete`
deployment would mean that the configuration would be cleared when
deploying, making the app fragile to a restart.

The app service doesn't change which version it's running unless it's
restarted.

This situation might be improved by setting the main slot to use the
currently live version explicitly, so configuration would match what's
already live, though this wouldn't stop environment variable changes
getting deployed. We could do this by querying the Azure API. I don't
believe this is a good idea, however. It's better to deploy a complete
configuration of the newest version accidentally than to deploy mixed
configuration.