Introduce Claim::VerifyResponseParametersParser
This will handle translating a Verify::Response into the attributes we
will use to update a claim.

Let's start with extracting gender. We don't need to worry about the
gender being "verified" and we're only interested in recording a value
of male or female as that is all that HMRC will accept for RTI payroll.