VerifyResponseParametersParser parses names
This parses name fields, mimicking the behaviour of Verify::Response.
The individual parts of the name are broken out to make it easier to
switch to capturing the name as separate attributes in the future.

This also fixes a bug where a value that didn't have a "from" date would
cause the response parser to fall over.