Back fill first, middle and surnames using Verify response
The gem "data_migrate" allows us to write data migrations in a similar
way we would with schema migrations. We can run them at the same time
or individually.

Since we only want to back fill claims that have had a successful Verify
response we scope our migration to only target those.

For now we're going to leave full_name intact and remove it later.

The table "data_migrations" is automatically created when you run the
`rake db:migrate:with_data` command. It's frustrating that it doesn't
generate it's own data_migrations schema migration as developers are
likely to run `rails db:migrate` when pulling down new changes not
schema:load. Because of this I've created a schema migration for it
that will only run if that table doesn't exist. Otherwise I ran in to an
error where the gem creates the table first:

```
PG::DuplicateTable: ERROR:  relation "data_migrations" already exists
```

There is an open issue about this on the gem's repo and it looks like it
will be addressed soon:
https://github.com/ilyakatz/data-migrate/issues/105