Run data:migrate after db:migrate
The data_migrate gem rewrites the `db:migrate` task in it's version
`db:migrate:with_data` to sort the schema and data migrations in to date
order before running them so that you can make changes to the schema
after migrating data. Rewriting core Rails tasks isn't ideal and may get
out of step with latest.

Also we want to ensure that the data migration was successful before
removing columns/destroying data, therefore it makes sense to have these
as two commands and require and schema cleanup to happen as part of a
separate PR.