Ignore standardrb warnings generated by data_migrate gem
When running `rails data:migrate` the data_migrate will put a magic
comment at the top of the file to set the encoding to UTF-8. Standard
will throw two warnings for this comment, one that it isn't needed at
all (the default is UTF-8) and the fact it needs an empty line after the
comment.

Since we can't control the output of data:migrate we should ignore it
for it now and possibly open up an issue/PR upstream on the gem's repo.