Use each instead of find_each in claims export
Active Record's find_each cannot be ordered and the claims export must
be ordered.

The claim export is a temporary measure for private beta only so we
are not concerned about the load on the service that switching to each
will incur as the number of claims is limited.

Details about in the Rails docs:

https://edgeguides.rubyonrails.org/active_record_querying.html#retrieving-multiple-objects-in-batches