Service operators can download a CSV containing payroll data
The CSV needs to be in a format that our payroll provider (Cantium) can
ingest.

The payroll data should only contain approved claims, but the process to
approve claims hasn't been finalised yet, so that will come later. We're
just using submitted claims for now.

There are a few fields that don't have values. That's because we either
don't capture the data, can't infer it or just not sure what value to
put.

Some of the values are bespoke static values that don't make sense
(eg. "T" for has a student loan). I decided to put these values in the
translations file so we can be a bit more descriptive with the
translation key, it also links it with the bespoke headers that Cantium
expects.

One of the acceptance criteria is that the leading zeros on the sort
code and account number are preserved. I have confirmed they are
preserved in the CSV but it's worth noting that if a user opens the CSV
in Excel or similar the software will strip the leading zero unless the
user explicitly sets the formatting for that cell. I think this is ok as
Cantium will ingest the CSV and it shouldn't be opened and saved by a
user.