Use self-hosted Google Analytics code
To make this work, stop using the gtag.js, which is for Google Analytics
and Google Tag Manager and instead use just the analytics.js, see docs
for the differences:

https://developers.google.com/analytics/devguides/collection/gtagjs/
https://developers.google.com/analytics/devguides/collection/analyticsjs/

gtag.js loads https://www.google-analytics.com/analytics.js into the
page so we are back to using a third party script when using gtag.js
locally.

We are not using the extra features from Tag Manager and only require
Google Analytics.

The source of the local analytics.js is:

https://www.google-analytics.com/analytics.js

and was obtained on Wed Sep 4 2019.

Prettier has modified the code style, but it is still uglified.