Only fill in address lines that exist
From the example file I had misunderstood the address lines to be
mappings to town, county, postcode etc. It was pointed out that these
address lines should be filled out as they exist. For example: If a
person only has address line 1 and postcode, the post code should be on
'address line 2' of the CSV.

We can achieve that by compacting the known address values in to one
`address_lines` array. We will never have an `address_line_6` so we
leave that as nil.