Configure OmniAuth path_prefix so that failure path works
Passing `path_prefix` as an option doesn't set it correctly, causing the
failure path to have the "/admin" part missing. This means users who
fail to authenticate themselves on DfE Sign In return to a 404 page.

By configuring OmniAuth's path_prefix we can ensure the correct failure
path will be generated. We were (incorrectly) overriding this value in
our spec helper, setting it globally means we don't have to do that any
more.