Student loans: schools that closed before policy start are ineligible
Eventually we'll have a concept of current claim year and this date
should use that, however for now we're just concerned with this year's
claimants.

If a school closed before the policy started then it is considered
ineligible, this should help users in choosing incorrect schools.

If the school closed after the policy started then we consider it
eligible as they will probably have student loan repayments outstanding
to claim back.