Drop redundant full_names column from claims
We split the claimants' names out to separate fields in Release 004,
making this field now redundant.

This also updates the data migration that performed the backfill so that
it won't error if it is run in future by a new developer joining the
team. There is probably an argument for simply deleting the data
migration, but given it's the only one we have it at the moment it might
be useful to keep it around to serve as a reference for future adding
future data migrations.