Update flakey spec
This spec would sporadically fail depending on the order the specs were
run in. It's not clear the order would impact this, but the only obvious
thing that might be causing it is that the warning message appearing and
the actual timeout had the same value of 1 second, meaning they would
fire at the same time and potentially the page will have timed out
before the spec runner had time to click the "Continue session" button
that resets the session.