Move and change format of date of birth
In both the database of qualified teachers and the teacher pension
service, date of birth is show in DD/MM/YYYY format. To make the task
of matching these dates simpler, display in the same format in the admin
claim view.

When checking claims, service operators generally check the claimants
date of birth before their national insurance number, so date of birth
should be shown first.