Refactor claim school questions to be an employment
In preparation for the multiple schools journey we need to refactor
the attributes associated with a claimant's time at a claim school in to
a new relationship: `employments`. For now we just have one employment,
so we introduce a `current_employment` method that just returns the last
or new. We also delegate a lot of the methods on `eligibility` to this
`current_employment` object. These are just temporary while we work on
the multiple schools journey.

Co-authored-by: Robbie Paul <robbiepaul@me.com>