Fix spec flakiness related to clashing identifiers
There was a [failed build] where a factory tried to create a local
authority district with a "code" that was already in the database.
Instead of sampling from the list of eligible codes, we can use the
fixtures that we create specifically for eligible and inelible local
authorities and districts.

- [failing build](https://dfe-ssp.visualstudio.com/S118-Teacher-Payments-Service/_build/results?buildId=22918)