Simplify CSV tests
We shouldn't have to test the output of the CSV rows in two places,
all we care about when testing the CSV itself is that the CSV has
the right headers, and the we have the right number of rows, with
roughly the correct output. The actual output of the rows themselves
is only important when we're actually testing the row. This also means we
only need to update one test when we make changes to the content of a
CSV row