Introduce concept of a claim's "award amount"
When preparing a payroll run, we will want to provide the user with
information about the total value of the claims that the run will
contain.

Since the value of a claim depends on the policy that the claim is being
made against, introduce an abstraction that allows us to query a claim for
its amount without having to know the underlying policy.

Co-authored-by: Tekin <tekin@tekin.co.uk>