Add a controller method to start the Claim journey
This initializes a `new` Claim with the relevant Eligibility and renders the
first page of the journey (in this case `qts_year`). Also makes some tweaks
to the `qts_year` template to make sure it accepts a non-persisted claim
and also submits a form to the create action if that's the case.