Make current_claim always return a claim
This cleans up the controllers, so we don't have to initialize claims in other
places and makes the intenrion of `current_claim` clearer. We do have to
change `send_unstarted_claiments_to_the_start` to check if a claim is
persisted, rather than present, but I don't think this is too much of an issue,
as this is only run on specific actions.