Approve button is disabled when payroll gender is missing
Claims that don't have a payroll gender cannot be approved because our
payroll provider (and HMRC) have gender as a required field. It is
possible for users to submit a claim and answer "I don't know" to the
question we ask about payroll gender.

We need to prevent admins from being able to approve these claims and
later flag it to them as something to investigate.