Move claim-related controller code to concern
Rather than defining claim-specific methods in the base controller and
having them present in all controllers of the application, we only
include them in the places where a current claim is expected and being
worked with.

This is part of the work to prepare us for introducing Maths & Physics
by reducing the places in the codebase where a specific policy is
referenced, i.e. in this case the `StudentoLoans::Eligibility`. In the
context of `ClaimsController` we will know the policy because of the
`:policy` param we get as part of the routing, which means we'll be
able to modify the method to set an appropriate eligibility based on
the param.

The only other place where the `current_claim` is checked and modified
is the `Verify::AuthenticationsController`. This doesn't have routing
scoped by a `:polocy`, but it also doesn't serve requests for unstarted
claims so that shouldn't be a problem.