Change the payroll start and end dates
The start and end dates in the payroll export is the fictional period a
claimant was employed in order for DfE to pay them.

We did not want the start date to be before the 6th of a given month as
the tax year runs from the 6 April until the 5 May the following year.

By setting the start date to the second Monday in the month and the end
date to the following Sunday, we avoid this issue.