Populate payments' award_amount from their claims
A notable thing about this is that, where we previously allowed Rails to
automatically create the payment records simply by specifying a payroll
run's claims association, we now have to explicitly create the payment
record, so that we can populate the award_amount column.

We did briefly try using a before_create callback on Payment to
automatically populate this value, but this was a non-starter:

- the before_create happens after validation, so the payment fails validation
due to missing award_amount;

- a few team members suggested that it's not a very good practice to
populate columns in a before_create, in general.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>