Move PayrollRun.payrollable_claims to Claim scope
A small refactor, suggested by Tekin. When originally adding the
payrollable_claims class method we did consider making it a scope, and
decided not to, with a vague intention of preventing the Claim class
becoming too large.

But scopes are probably to be preferred, in general: they're composable
and can be used with associations.