Change claim decision to be radio buttons
It was felt that having a radio button instead of two seperate submit
buttons would make the act of approving/rejecting a claim more
purposeful.

We have a slight discrepancy between the language in our code and the
language in the UI. In the UI we're using "claim decision", in the code
we're using "claim check/result". We may want to refactor our code to
match the UI and make it easier to understand.