Operator can upload a Payment Confirmation Report
A Payment Confirmation Report is a CSV file issued by Cantium after they
have calculated the payroll figures for the claims in a payroll run.
This commit enables a service operator to upload a Payment Confirmation
Report for a payroll run. The figures from this report will be stored in
the payment records which link the payroll run to its claims.

The CSV examples used in the specs are based on the file that came from
the Cantium trial run known as "TEST 7".

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>