In specs, link claims to payroll runs via factory
This is a refactor that overall improves quality of life whilst writing
specs.

Previously, to set up a spec with a list of claims linked to a payroll
run, we had to create a list of claims and then manually create the
Payment instances. Now, we can pass claims_count to the payroll_run
factory, which will create a list of claims, and create payments linking
the claims to the payroll run.

In the specs where we choose to use this approach, we sacrifice the
flexibility to craft the attributes of an individual claim. The only
place where we were doing this at the moment is in claims_csv_spec.rb,
for which this loss is no real issue. Something to bear in mind for the
future, though.

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>