Create Payment model
Payment acts as a join model which links a claim to a payroll run. We
are making this change so that we can attach additional payment-related
information to the payment, such as tax, national insurance, etc.

The first additional column that we introduce is award_amount, which
will act as an intentional de-normalisation of the data stored in the
claims' validities, as currently accessed via the polymorphic
Claim#award_amount method. We are doing this because we _anticipate_
that it will be useful to be able to access these amounts using a single
table and a single column. This commit does not populate the
award_amount column; this will come in a later commit.

This commit includes the data migration to create payments for existing
claims which are currently associated with a payroll run via the
payroll_run_id column. We will need to remove this column at some point.

Data migration spec approach copied from
https://github.com/DFE-Digital/dfe-teachers-payment-service/blob/3c1ca10ccb8df22a0696966bddb1e8caf62a70a0/spec/db/data/populate_student_loans_employments_spec.rb

Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>