Don't ask for specific year for QTS year
We did originally think we'd need the precise QTS year from the teacher
so that we could use it to estimate their claim amount based on their
likely salary. But now that we have access to TPS and use it for claim
checking, we have more accurate data for their salary.
Therefor we no longer need the precise year and can simplify the
question to "before" or "after" the eligibility date.