Add payroll runs and claims to development seeds
All the data generated by the factories is constant (e.g. claimants'
names) so it's not the most realistic seed data in the world, but it
gives a useful set of data out of the box that allows a developer to
test / demo all of the admin functionality.

We might want to consider using a gem like Faker to enable the factories
to generate random realistic data (names, addresses, emails, ...)