Introduce a MatchingAttributeFinder for claims
Finds other submitted claims that have details that match those used in the
source claim.

We know that finding matched details in claims is an expensive operation. We
looked at ways to optimise the query and found that due to its nature, the query
was always run seqentially so indexes on the columns did not improve the
performance.

The class returns a hash of claim ids and and an array of matching attributes
names for each claim for displaying to the service operator alongside the
source claim.

PR changes:
Co-authored-by: pezholio <pezholio@gmail.com>
Co-authored-by: Lawrence Forooghian <lawrence@dxw.com>