Introduce generic Claim::PermittedParameters class
This new class for determining the attributes a user should be able to
edit whilst making a claim is policy-agnostic, meaning it will be
possible to use it when the Maths & Physics policy is introduced: all
that will be required is for a
MathsAndPhysics::Eligibility::EDITABLE_ATTRIBUTES array to be defined.

This is part of the work to lay the ground work for the second policy to
the service.