Remove ClaimUpdate
ClaimUpdate has now been refactored to the point where its existence
isn't really justified and it's just as clear, and with less indirection
if the action of resetting dependent answers is performed inline in the
controller action. Some extra test coverage on the controller action has
been added for extra safety.