Create mailer action to send payment notification
Creating a new mailer action to send a payment confirmation email to a user.
The payment date needs to be sent as a Unix timestamp (old skool huh?),
because `ActiveJob` can't serialize Date objects properly. I could have
calculated the payment date in the mailer, but I don't like to have too much
logic in mailers, and as the mailers run asynchronously, we could end up with
unpredictable results if the mailers get stuck in a queue for some reason.