Enqueue emails after a successful ingest
Once a successful ingest has taken place, we then queue up all the
confirmation emails to be sent to users. Because ingests will usually
performed on a Tuesday, we put our faith in the process, and tell the user
the payment date will be on the Friday of that current week, though this may
evolve as time goes on.