Load GOVUK Frontend from npm package
Loading the GOVUK Frontend styles, js and assets from a gem worked great until
it didn't when an update to sprockets broke it.

Althought the gem is fixed, as the GOVUK Frontend is distrubuted as a npm
package so it makes more sense to load it from there and keep the dependency
tighter.

This commit removes the gem and makes the changes to load the assets from the
npm package in node_modules.