Make it clear we may deduct student loan repayment
The current copy makes it sound like the claimant will always receive
the full amount that they claimed. This is not true, as we may need to
deduct a student loan repayment from the amount they receive.

Also, tweak the copy a little for readability and consistency.

These changes come from Faye, our content designer.