Move static pages under policy in URL
In preparation for introducing Maths & Physics, this moves the static
pages so that they sit under the policy root in the URL path, e.g.
"/privacy" moves to "/student-loans/privacy". This will allow us to
customise these pages based on the specific policy, for example with an
appropriate support email address or the main page heading.