Move current school closed check to school checker
Now that we have a place where all the other current school eligibility
checks are being done, it makes sense to move the check of whether the
current school is open.

This change was suggested by Robbie and Tekin in a code review.

This is not a pure refactor; it is a slight regression in functionality
since we lose the specific ineligbility message. However, looking at the
original pull request in which this check was introduced, it is not
something we anticipate happening in real life, since the current school
search only displays schools which are open. We decided we are happy
with this loss of functionality, in return for better encapsulation.