Get rid of temporary start page and redirect root
Once we're live for public beta, the start page will be hosted elsewhere
so any visitors to the root of the domain should be redirected there,
rather than our temporary start page. This was the last place we were
referencing the temporary start page so it's now safe to remove it.