Make session refresh claim-agnostic
The action that refreshes the user's claim session doesn't need to be
namespaced to the policy as all it does is respond with an :ok response
after refreshing the active session. This gets rid of one of the two
routes that are hardcoded to the "student-loans" policy from
timeout_dialogue.js.erb, paving the way for making it multi-policy
compatible.

Making session refresh a top-level route also makes sense because the
same JS and action is used for the admin part of the service, which
shouldn't be coupled to particular policy's routing.