Create current_policy_routing_name helper
This makes the application more robust in that we can use different
methods to determine the current policy path, based on the context.
For example there is no `:policy` param for non-namespaced routes such
as the GOV.UK Verify callback ones. For those, which are part of the
claim journey, we can override the method and determine the policy path
using the current_claim.