Allow user to restart claim by visiting start page
If the user submits answers that make their claim ineligible, we want
them to still be able to start a new claim.

This is fixing a regression, that was introduced when we removed our
temporary start page, which performed a POST to the first question.

As a future piece of work, we should also consider allowing the user to go
back to the answer that caused their claim to become ineligible, instead
of having to restart their claim.