Make timeout dialog JS work for admin users
Up to now the JavaScript-based timeout dialog and reset would not have
worked unless the admin user had an active claim session in progress.
This drops the "claim-in-progress" data attribute and instead runs the
logic for the timeout dialog based on the presence of the dialog box
instead.

I've copied the testing pattern we're using for the public-facing
timeout. I'm not entirely happy with the way these specs rely on
stubbing out internal methods on the controller code, but synchronising
time manipulation between the server running the app code and the
client-side JavaScript is hard. Something I'd like us to look into in
the future.