Move admin session code to admin controller
This mostly organises the code for admin sessions and public-facing user
sessions such that they are entirely separate, the exception being the
update_last_seen_at method, which updates the timestamp for the user
regardless of the type of user they are.

A test was deleted as part of this mostly because it was a mostly false
requirement - we do not expect admin users to be visiting the
public-facing site, and if they do, it's not a major issue that their
sign-in session will get extended because of it.