Remove default "student-loans" policy from routing
A lot of the previous Maths & Physics preparation work was leading to
this moment where we can remove the default policy in the routing and
be explicit about the current active `:policy` in the app code. Now it
will be possible to introduce additional policy routing without breaking
any of the existing views, controllers or spec code.

To keep the rendering of the maintenace page working,
current_policy_routing_name had to be overriden in StaticPagesController
as the catch-all route for maintenance mode won't include a value for
:policy in the params.

Note that at the moment there is little to distinguish the specs that
have to be coupled to the policy, for example
spec/features/student_loans_claim_spec.rb, and those that shoud be more
generic as they cover behaviour that is policy-agnostic, such as
spec/requests/claims_spec.rb. At some point very soon these specs should
be rationalised to make it obvious which are testing behaviour specific
to a given policy, and which are policy-agnostic.