Parameterise routes to prepare for Maths & Physics
This makes it such that introducing routing for Maths & Physics will
simply be a case of defining a `MathsAndPhysics::SlugSequence` and
adding the `MathsAndPhysics` module to the array of `policies`.

I've made use of a routing constraint object here as I struggled to get
the routing to work just using the default constraints and defaults when
a second policy was introduced.