Upcase when normalising National Insurance number
Now that we're performing case-insensitive comparisons when detecting
matching claims, this is no longer strictly necessary. It does however
seem like good data hygiene, and something that we should have done when
we first wrote the normalisation.

There will also be a data migration to upcase the existing NI numbers,
but this will be deployed in a subsequent release, to ensure that,
before we run the data migration, all of the running web servers are
already using this new normalisation. Since we have zero-downtime
deploys, we cannot make this guarantee if we deploy the new code and the
data migration simultaneously.