Fix flakey current school search spec
We're not sure what's causing the error to happen, some combination of
headless chrome not liking the `<p class="govuk-body">` tag. Changing
the paragraph's class to anything else fixes the error. Adding a sleep
before the failing expectation also fixes the error.

We felt that this solution was smallest change that we could get to pass
consistently.

We had concerns about the way that screen readers might read the two
span tags that are back to back, but we checked it in VoiceOver and it
made sense.