Try another school if school is ineligible
This allows a user to go back into the flow and try another school if they find
their initial claim school is ineligible (for example, they worked in more than
one school during that period).

When a user clicks "Enter another school" a parameter is set that alters the
title to make more sense in the context of the flow, as well as setting the
`school_search_value` to nil, which means the previous school won't be
prepopulated.

If a user has exhausted all their options, they are then sent to the ineligible
page again, but this time with a parameter that means they see a more
generic (and final) ineligiblity message.