Policy-unaware controller is able to redirect user
Because the routes for the verify actions are not namespaced to a
particular policy it won't be possible for the controller to redirect
users to a start page if they visit the URLs without an active claim. We
can redirect these requests to the root, which will be handled
appropriately, either as it is now by redirecting to the Student Loans
start page, or in the future when we introduce a page for lost users
pointing at both policies once Maths & Physics is introduced.