Make PartOfClaimJourney policy-aware
To now PartOfClaimJourney was coupled to the StudentLoans policy. This
updates it so that it will correctly initialise new claims and redirect
unstarted claimants according to the actual policy defined by the
routing parameters. Note that current_claim needs to be able to behave
even when there is no `params[:policy]` because
Verify::AuthenticationsController is not namespaced by policy.