Create PolicyConfiguration model
This will store per-policy configuration for the application. The first
piece of configuration will enable us to build the admin interface that
allows a service operator to open or close a policy for submissions
(which will come in a future commit).

The application expects that a policy configuration record will be
present in the database for each policy. Because these records are
fundamental to the operation of the app, I have added Rails test
fixtures for them, which I think is in line with our current usage of
fixtures.

The records will be added to the existing production database by the
data migration, and they would be seeded into a fresh production or
development database by db/seeds.rb.