Prevent claiming if policy closed for submissions
Use a policy's configuration record to determine whether to display the
"policy closed" page.

A future commit will add an admin UI to change the policies'
configurations.

Note that this contains an additional functional change: we allow the
user to continue to access the static pages such as privacy notice and
accessibility statements, even when that policy is closed for
submissions.