Move mailer preview to spec folder
Because autoloading is handled differently in Rails 6, everything in the `app/`
folder is autoloaded. This is fine in development, because
`config.action_mailer.preview_path` is set, and the file is evaluated when the
app boots, but in production, the file is not evaluated, so Rails raises a
`Zeitwerk::NameError`, because it doesn't follow the underlying rules for
autoloading in Rails.

Moving the mailer preview to the default location works because the `spec`
folder is not evaluated by rails on load.