A service operator can create a DQT report request
This uses the new Claim::DatabaseOfQualifiedTeachersReportRequest class to
export the claims used to create a report from the Database of qualified
teachers as a csv.

When asked, service operators felt the 'view claims' screen was the right place
to find this feature.

We do not expect this feature to remain in the service once we have a connection
to the Database of qualified teachers.