Show the payroll run file due date
Service operators should be shown when the next date the payroll file is due
to Cantium.

If no payroll run has happend yet this month, this date is the third to last
Friday of the current month.

If a payroll has already happened this month, it will be the the third to last
Friday of the next month.