Remove default maintenance availability message
Whilst designing the UI for the admin maintenance inferface, we
discussed whether we should include the default message in the hint text
for the availability message textarea, so that service operators will
know what happens if they enter a blank message.

But then, we decided that the default message probably isn't necessary
at all. There's no reason to believe that "You will be able to use the
service later today" is a particularly representative example of
maintenance availability.