Reinstate production basic auth env vars
These were hardcoded to blank values to disable basic auth when we went
into public beta. Instead of doing that in the template we should load
the value from the key vault, that way we can enable/disable basic auth
without having to do a full release.

We're currently making use of basic auth to protect the new Maths &
Physics journey.